Reorganised tests to separate the delegate pattern tests
