+++ src/tests/tests.csproj
@@ -12,6 +12,7 @@
 
     <ItemGroup>
         <PackageReference Include="JUnitTestLogger" Version="1.1.0" />
+        <PackageReference Include="Moq" Version="4.20.70" />
         <PackageReference Include="nunit" Version="4.1.0" />
         <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0-release-24177-07" />
